Add data checks to prevent ckmeans/simple statistics errors
